removal of old files
