+++ app/src/main/java/de/tum/in/tumcampusapp/service/GeofencingStartupReceiver.kt
@@ -10,19 +10,16 @@ import de.tum.`in`.tumcampusapp.utils.Utils
 class GeofencingStartupReceiver : BroadcastReceiver() {
 
     override fun onReceive(context: Context?, intent: Intent?) {
+        Utils.log("HOSSMARK Received geofencing broadcast")
         if (!isValidIntent(intent)) {
             return
         }
 
         Utils.log("Restarting geofencing due to " + intent?.action)
         context?.let {
-            /*val geofencingIntent = GeofencingRegistrationService.buildGeofence(it, MUNICH_GEOFENCE,
+            val geofencingWorker = GeoFencingRegistrationService.buildGeofence(it, MUNICH_GEOFENCE,
                     48.137430, 11.575490, DISTANCE_IN_METER)
-            GeofencingRegistrationService.startGeofencing(it, geofencingIntent)*/
-
-            val geofencingWorker = GeoFencingRegistrationServiceRW.buildGeofence(it, MUNICH_GEOFENCE,
-                    48.137430, 11.575490, DISTANCE_IN_METER)
-            GeoFencingRegistrationServiceRW.startGeofencing(it, geofencingWorker)
+            GeoFencingRegistrationService.startGeofencing(it, geofencingWorker)
 
         }
     }
