+++ app/src/main/java/de/tum/in/tumcampusapp/service/StartSyncReceiver.kt
@@ -23,8 +23,7 @@ class StartSyncReceiver : BroadcastReceiver() {
 
         // Also start the SilenceService. It checks if it is enabled, so we don't need to
         // SilenceService also needs accurate timings, so we can't use WorkManager
-        //SilenceService.enqueueWork(context, Intent())
-        SilenceServiceRW.enqueueWork(context)
+        SilenceService.enqueueWork(context)
     }
 
     companion object {
