+++ app/src/main/java/de/tum/in/tumcampusapp/utils/CacheManager.kt
@@ -5,8 +5,7 @@ import de.tum.`in`.tumcampusapp.api.tumonline.CacheControl
 import de.tum.`in`.tumcampusapp.api.tumonline.TUMOnlineClient
 import de.tum.`in`.tumcampusapp.component.tumui.calendar.CalendarController
 import de.tum.`in`.tumcampusapp.component.tumui.calendar.model.EventsResponse
-import de.tum.`in`.tumcampusapp.service.QueryLocationServiceRW
-import de.tum.`in`.tumcampusapp.service.QueryLocationsService
+import de.tum.`in`.tumcampusapp.service.QueryLocationService
 import okhttp3.Cache
 import org.jetbrains.anko.doAsync
 import retrofit2.Call
@@ -45,8 +44,7 @@ class CacheManager @Inject constructor(private val context: Context) {
 
     private fun loadRoomLocations() {
         doAsync {
-            //QueryLocationsService.enqueueWork(context)
-            QueryLocationServiceRW.enqueueWork(context)
+            QueryLocationService.enqueueWork(context)
         }
     }
